Image upload check (#75)
* Add context

* Add check for file type

* Add check for file size

* Update tests

* Resolve method for checking file size

* Resolve error handling and test

* Update src/App/pinJSONAndImage/pin.middleware.ts

File size review


* Update return type

* Make sure file is deleted on error

+++ src/App/pinJSONAndImage/pin.resolver.ts
@@ -2,7 +2,12 @@ import { Args, Mutation, Resolver } from '@nestjs/graphql'
 import { GraphQLUpload, FileUpload } from 'graphql-upload'
 import { createWriteStream } from 'fs'
 import * as fs from 'fs/promises'
-import { InternalServerErrorException, Logger } from '@nestjs/common'
+import {
+  HttpException,
+  HttpStatus,
+  InternalServerErrorException,
+  Logger,
+} from '@nestjs/common'
 import IpfsHash from './model/ipfsHash'
 import PinService from './pin.service'
 
@@ -19,23 +24,39 @@ class PinResolver {
     }
   }
 
+  private async checkFile(image: FileUpload) {
+    const type = image.mimetype.split('/')[0]
+    if (type !== 'image') {
+      Logger.error('Wrong file type')
+      throw new HttpException(
+        'Unsupported file type! Make sure file is an image.',
+        HttpStatus.BAD_REQUEST,
+      )
+    }
+    return image
+  }
+
   @Mutation(() => IpfsHash, { name: 'pinImage' })
   async pinImage(
     @Args({ name: 'fileUpload', type: () => GraphQLUpload })
     image: FileUpload,
   ): Promise<FileUpload> {
-    const { createReadStream, filename } = await image
+    const { createReadStream, filename } = await this.checkFile(image)
     const destinationPath = `./uploads/${filename}`
-    return new Promise((res, rej) =>
-      createReadStream()
-        .pipe(createWriteStream(destinationPath))
-        .on('error', rej)
-        .on('finish', async () => {
-          const result = await this.pinService.pinImage(destinationPath)
+    return new Promise((res, rej) => {
+      const read = createReadStream()
+      read.on('error', async err => {
+        await fs.unlink(destinationPath)
+        rej(err)
+      })
+      read.pipe(createWriteStream(destinationPath)).on('finish', async () => {
+        const result = await this.pinService.pinImage(destinationPath)
+        if (result) {
           await fs.unlink(destinationPath)
-          res(this.errorHandler(result))
-        }),
-    )
+        }
+        res(this.errorHandler(result))
+      })
+    })
   }
 
   @Mutation(() => IpfsHash, { name: 'pinJSON' })
